Move systemError.js to core/
