+++ server/utils/unauthorizedError.js
@@ -1,4 +1,4 @@
-const SystemError = require('./systemError')
+const SystemError = require('@core/systemError')
 
 class UnauthorizedError extends SystemError {
   constructor (userName) {
